+++ common/src/main/java/immersive_aircraft/network/ClientNetworkManager.java
@@ -44,9 +44,6 @@ public class ClientNetworkManager implements NetworkManager {
         ClientLevel level = Minecraft.getInstance().level;
 
         if (level != null) {
-            // Sound
-            level.playLocalSound(fireResponse.x, fireResponse.y, fireResponse.z, SoundEvents.CROSSBOW_SHOOT, SoundSource.NEUTRAL, 1.0f, 1.0f, true);
-
             // Particles
             RandomSource random = level.getRandom();
             double r = 0.1;